Revert "feat: updated protobufs, added token pause functionality (#288)" (#295)
This reverts commit 951773cefa14632281e1bb6dc1470ce11fd164e3.

+++ account_create_transaction.go
@@ -160,7 +160,7 @@ func (transaction *AccountCreateTransaction) _Build() *proto.TransactionBody {
 		ReceiveRecordThreshold:        transaction.sendRecordThreshold,
 		ReceiverSigRequired:           transaction.receiverSignatureRequired,
 		Memo:                          transaction.memo,
-		MaxAutomaticTokenAssociations: int32(transaction.maxAutomaticTokenAssociations),
+		MaxAutomaticTokenAssociations: transaction.maxAutomaticTokenAssociations,
 	}
 
 	if transaction.key != nil {
